requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
bower==0.0.0
Django==3.2.5
django-apscheduler==0.6.0
django-crispy-forms==1.12.0
django-python3-ldap==0.12.0
ldap3==2.9.1
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-ldap==3.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1

djangorestframework~=3.13.1